数据输入/tools/candidate_sync/requirements.txt
pandas>=2.0.0
pyarrow>=14.0.2
python-dateutil>=2.8.2
pyyaml>=6.0.1
python-dotenv>=1.0.1

